Improving the forwarding of logged-in users.
When forwarding logged-in users from some path to a dashboard, we now have a loading state to avoid temporary flash of that path's contents. SecureRoute handles this.

Also introducing loggedOut redux store state. This is in order to know when we should not have loading state (avoiding infinite loading state after logout).