Logging in with popup in progress
Able to sign in with Google with popup. The same user is able to both register/login with both email/password and google. However, see the TODO in firebaseModel listener middleware in order to avoid overwriting existing user data when registering/logging in with Google (after having created user data (while logged in with email/password or with google).